// backend/src/modules/tasks/task.dto.ts
import { z } from "zod";

export const CreateTaskDto = z.object({
    title: z.string().max(100),
    description: z.string(),
    dueDate: z.string(),
    priority: z.enum(["LOW", "MEDIUM", "HIGH", "URGENT"]),
    status: z.enum(["TODO", "IN_PROGRESS", "REVIEW", "COMPLETED"]),
    assignedToId: z.string().optional(),
});

export const UpdateTaskDto = z.object({
    title: z.string().max(100).optional(),
    description: z.string().optional(),
    dueDate: z.string().optional(),
    priority: z.enum(["LOW", "MEDIUM", "HIGH", "URGENT"]).optional(),
    status: z.enum(["TODO", "IN_PROGRESS", "REVIEW", "COMPLETED"]).optional(),
    assignedToId: z.string().optional(),
});

export type CreateTaskInput = z.infer<typeof CreateTaskDto>;
export type UpdateTaskInput = z.infer<typeof UpdateTaskDto>;